# Add --perf option