--- a/include/myst/options.h
+++ b/include/myst/options.h
@@ -17,6 +17,7 @@ typedef struct myst_options
     bool shell_mode;
     bool debug_symbols;
     bool memcheck;
+    bool perf;
     bool report_native_tids;
     size_t max_affinity_cpus;
     char rootfs[PATH_MAX];