--- a/include/myst/times.h
+++ b/include/myst/times.h
@@ -6,14 +6,16 @@
 
 #include <time.h>
 
+long myst_lapsed_nsecs(const struct timespec* t0, const struct timespec* t1);
+
 /* Start tracking time for current thread */
 void myst_times_start();
 
 /* Time tracking while entering the kernel from user space */
-void myst_times_enter_kernel();
+void myst_times_enter_kernel(long syscall_num);
 
 /* Time tracking while leaving the kernel to user space */
-void myst_times_leave_kernel();
+void myst_times_leave_kernel(long syscall_num);
 
 /* Return the time (in nanoseconds) spent on kernel execution */
 long myst_times_system_time();
@@ -26,6 +28,8 @@ long myst_times_process_time();
 
 /* Return the time (in nanoseconds) spent by thread */
 long myst_times_thread_time();
+
+void myst_print_syscall_times(const char* message, size_t count);
 
 /* Return nanoseconds since startup */
 long myst_times_uptime();