--- a/solutions/aspnet/Program.cs
+++ b/solutions/aspnet/Program.cs
@@ -21,7 +21,7 @@ namespace webapp
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseStartup<Startup>();
-                    webBuilder.UseUrls("http://0.0.0.0:5050");
+                    webBuilder.UseUrls("http://127.0.0.1:5050");
                 });
     }
 }