--- a/solutions/python_webserver/Makefile
+++ b/solutions/python_webserver/Makefile
@@ -3,16 +3,33 @@ include $(TOP)/defs.mak
 
 APPBUILDER=$(TOP)/scripts/appbuilder
 
+OPTS =
+
 ifdef STRACE
-OPTS = --strace
+OPTS += --strace
 endif
 
-OPTS = --memory-size=1024m
+OPTS += --memory-size=256m
+
+ifdef PERF
+OPTS += --perf
+endif
 
 all: rootfs
 
+USE_ROOTHASH=1
+
 rootfs: appdir
-	$(MYST) mkcpio appdir rootfs
+	sudo $(MYST) mkext2 appdir rootfs
+ifdef USE_ROOTHASH
+	$(MYST) fssig --roothash rootfs > roothash
+else
+	truncate --size=-4096 rootfs
+endif
+
+ifdef USE_ROOTHASH
+OPTS += --roothash=roothash
+endif
 
 appdir:
 	$(APPBUILDER) Dockerfile
@@ -21,10 +38,10 @@ TIMEOUT=60s
 
 run:
 	test -f server.pid && kill -9 `cat server.pid` || true
-	$(MYST_EXEC) $(OPTS) rootfs /miniconda/bin/python3 /app/hello_server.py > server.output & echo $$! > server.pid
-	timeout $(TIMEOUT) tail -f server.output | ./client.sh
+	$(MYST_EXEC) $(OPTS) rootfs /miniconda/bin/python3 /app/hello_server.py & echo $$! > server.pid
+	./wait.sh
+	curl 127.0.0.1:8000
 	test -f server.pid && kill -9 `cat server.pid` && rm server.pid || true
-	test -f client.output
 
 clean:
 	test -f server.pid && kill -9 `cat server.pid` || true