--- a/solutions/python_webserver/hello_server.py
+++ b/solutions/python_webserver/hello_server.py
@@ -16,7 +16,7 @@ class MyWebServer(BaseHTTPRequestHandler):
         self.wfile.write(bytes("\n", "utf-8"))
 
 def run(server_class=HTTPServer, handler_class=MyWebServer):
-    server_address = ("0.0.0.0", 8000)
+    server_address = ("127.0.0.1", 8000)
     httpd = server_class(server_address, handler_class)
     print("launching server...")
     httpd.serve_forever()