--- a/solutions/python_webserver/wait.sh
+++ b/solutions/python_webserver/wait.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+while [ -z "${str}" ]
+do
+    str=$(sudo lsof -i -P -n | grep "\<8000\>")
+done