--- a/tools/myst/myst.edl
+++ b/tools/myst/myst.edl
@@ -81,7 +81,9 @@ enclave
             [in, size=mount_mappings_size] const void* mount_mappings,
             size_t mount_mappings_size,
             uint64_t event,
-            pid_t target_tid);
+            pid_t target_tid,
+            uint64_t start_time_sec,
+            uint64_t start_time_nsec);
 
         public long myst_run_thread_ecall(
             uint64_t cookie,