… solutions/python_webserver/client.sh