This PR adds the ``myst --perf`` option that prints a report indicating how much time is spent in each syscall. For example.

```
$ myst exec-sgx --perf rootfs /bin/getcwd

=== boot time: 1.3535sec
=== passed test (/bin/getcwd)

============================================================
kernel shutdown: 1.3720 seconds elapsed
============================================================
SYS_mmap                   0.0005sec 33.52%  8.38% (4 calls)
SYS_exit_group             0.0001sec  7.44%  7.44% (1 calls)
SYS_myst_is_shared_crt     0.0001sec  7.43%  7.43% (1 calls)
SYS_mprotect               0.0001sec  7.37%  3.69% (2 calls)
SYS_set_thread_area        0.0001sec  7.36%  7.36% (1 calls)
SYS_open                   0.0001sec  7.33%  7.33% (1 calls)
SYS_writev                 0.0001sec  7.31%  7.31% (1 calls)
SYS_myst_trace             0.0001sec  3.82%  3.82% (1 calls)
SYS_close                  0.0001sec  3.70%  3.70% (1 calls)
SYS_ioctl                  0.0001sec  3.69%  3.69% (1 calls)
SYS_getcwd                 0.0001sec  3.69%  3.69% (1 calls)
SYS_read                   0.0001sec  3.68%  3.68% (1 calls)
SYS_myst_add_symbol_file   0.0001sec  3.66%  3.66% (1 calls)
SYS_set_tid_address        0.0000sec  0.00%  0.00% (1 calls)
SYS_myst_load_symbols      0.0000sec  0.00%  0.00% (1 calls)
```